requirements.txt
Flask
Flask-PyMongo
email-validator
face-recognition
opencv-python
Pillow
python-dotenv
Werkzeug